+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - The following methods for `RArray`: `dup`, `concat`, `plus`, `delete`,
   `delete_at`, `resize`, `reverse`, `rotate, and `sort`.
 - New methods for `RString`: `enc_new`, `len`, `length`, and `is_empty`.
+- `RHash` gains the methods `delete` and `clear`.
 
 ### Changed
 - `RArray::cat`, `RArray::from_slice`, and `gc::mark_slice` will accept a